Revert "General 2D Reduction Kernel (#2535)"
This reverts commit 4750b293fe0abfa44a32181742a48b1dfec468f7.
