+++ include/ck_tile/core/utility/reduce_operator.hpp
@@ -26,8 +26,7 @@ struct Add
     }
 
     template <typename T,
-              typename = std::enable_if_t<std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
+              typename = std::enable_if_t<std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t>>>
     CK_TILE_HOST_DEVICE constexpr T operator()(T& y, T x) const
     {
         float y_ = type_convert<float>(y);
@@ -35,8 +34,6 @@ struct Add
 
         return type_convert<T>(y_ + x_);
     }
-
-    static constexpr bool requires_special_combine = false;
 };
 
 struct SquareAdd
@@ -54,47 +51,13 @@ struct SquareAdd
     {
         return y + (x * x);
     }
-
-    template <typename T,
-              typename = std::enable_if_t<std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
-    CK_TILE_HOST_DEVICE constexpr T operator()(T& y, T x) const
-    {
-        float y_ = type_convert<float>(y);
-        float x_ = type_convert<float>(x);
-        return type_convert<T>(y_ + (x_ * x_));
-    }
-
-    // For combining partial results
-    template <typename T,
-              typename = std::enable_if_t<std::is_same_v<T, float> || std::is_same_v<T, double> ||
-                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t>>>
-    CK_TILE_HOST_DEVICE constexpr T combine_partial_results(const T& partial1,
-                                                            const T& partial2) const
-    {
-        return partial1 + partial2; // Just add the partial sums, don't square again
-    }
-
-    template <typename T,
-              typename = std::enable_if_t<std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
-    CK_TILE_HOST_DEVICE constexpr T combine_partial_results(T& partial1, T& partial2) const
-    {
-        float partial1_ = type_convert<float>(partial1);
-        float partial2_ = type_convert<float>(partial2);
-        return type_convert<T>(partial1_ + partial2_);
-    }
-
-    static constexpr bool requires_special_combine = true;
 };
 
 struct Max
 {
     template <typename T,
               typename = std::enable_if_t<std::is_same_v<T, float> || std::is_same_v<T, double> ||
-                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t> ||
-                                          std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
+                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t>>>
     CK_TILE_HOST_DEVICE static constexpr T GetIdentityValue()
     {
         return numeric<T>::min();
@@ -102,24 +65,18 @@ struct Max
 
     template <typename T,
               typename = std::enable_if_t<std::is_same_v<T, float> || std::is_same_v<T, double> ||
-                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t> ||
-                                          std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
+                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t>>>
     CK_TILE_HOST_DEVICE constexpr T operator()(const T& y, const T x) const
     {
         return max(y, x);
     }
-
-    static constexpr bool requires_special_combine = false;
 };
 
 struct AbsMax
 {
     template <typename T,
               typename = std::enable_if_t<std::is_same_v<T, float> || std::is_same_v<T, double> ||
-                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t> ||
-                                          std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
+                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t>>>
     CK_TILE_HOST_DEVICE static constexpr T GetIdentityValue()
     {
         return numeric<T>::min();
@@ -127,15 +84,11 @@ struct AbsMax
 
     template <typename T,
               typename = std::enable_if_t<std::is_same_v<T, float> || std::is_same_v<T, double> ||
-                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t> ||
-                                          std::is_same_v<T, half_t> || std::is_same_v<T, bf16_t> ||
-                                          std::is_same_v<T, fp8_t> || std::is_same_v<T, bf8_t>>>
+                                          std::is_same_v<T, int32_t> || std::is_same_v<T, int8_t>>>
     CK_TILE_HOST_DEVICE constexpr T operator()(const T& y, const T x) const
     {
         return max(y, abs(x));
     }
-
-    static constexpr bool requires_special_combine = false;
 };
 
 } // namespace ReduceOp
