+++ include/ck_tile/ops/reduce.hpp
@@ -5,11 +5,8 @@
 
 #include "ck_tile/ops/reduce/block/block_reduce.hpp"
 #include "ck_tile/ops/reduce/block/block_reduce2d.hpp"
+#include "ck_tile/ops/reduce/block/block_reduce2d_default_policy.hpp"
 #include "ck_tile/ops/reduce/block/block_reduce2d_problem.hpp"
 #include "ck_tile/ops/common/generic_2d_block_shape.hpp"
 #include "ck_tile/ops/common/tensor_layout.hpp"
 #include "ck_tile/ops/common/utils.hpp"
-#include "ck_tile/ops/reduce/kernel/reduce2d_kernel.hpp"
-#include "ck_tile/ops/reduce/pipeline/reduce2d_default_policy.hpp"
-#include "ck_tile/ops/reduce/pipeline/reduce2d_problem.hpp"
-#include "ck_tile/ops/reduce/pipeline/reduce2d_shape.hpp"
