+++ test/ck_tile/CMakeLists.txt
@@ -22,4 +22,3 @@ add_subdirectory(add_rmsnorm2d_rdquant)
 # add_subdirectory(rmsnorm2d)
 add_subdirectory(gemm_block_scale)
 add_subdirectory(utility)
-add_subdirectory(reduce)
