--- test/ck_tile/reduce/CMakeLists.txt
@@ -1,7 +1,0 @@
-if(GPU_TARGETS MATCHES "gfx9")
-    add_gtest_executable(test_ck_tile_reduce2d test_reduce2d.cpp)
-    if(result EQUAL 0)
-        target_link_libraries(test_ck_tile_reduce2d PRIVATE utility)
-    endif()
-endif()
-
